Output a new meta element containing the source of the news item
This allows to style the items with CSS depending on the source. In
practice, I use it to hide the ads in the Slashdot feed with this rule:

    .source-slashdot p:not(.blosxomEntryFoot),
    .source-slashdot .share_submission {
        display: none;
    }

The :not() selector is defined in CSS3. It is supported in all modern
browsers, i.e. IE9+ and all the other ones.

Also see http://quirksmode.org/css/contents.html